requirements.txt
Django==5.1.1
djangorestframework==3.15.2
djangorestframework_simplejwt==5.3.1
gunicorn==23.0.0
psycopg2-binary==2.9.10
